refactor: App の push helper 群を Relay::PushHelpers へ抽出する (#27)
app.rb の helpers ブロックのうち push 送信・結果ハンドリング系 11 メソッド
（build_push_payload / dispatch_push / push_client_for / log_push_received /
handle_push_result / push_context / handle_push_delivered / handle_wns_status /
handle_push_gone / handle_push_oversized / handle_push_failed）と WNS_BENIGN_STATUSES
定数を lib/relay/push_helpers.rb の Relay::PushHelpers へ移し、App では
`helpers Relay::PushHelpers` で mixin する。ロジックの書き換えは無い純粋な移動。
全 route が使う authenticate! / json_body は App に残す。

Sinatra の helpers はモジュール mixin なので settings / request / halt / status の
request scope はそのまま効く。App は config/settings.yml を読む configure ブロックが
あり test から require できないため、push_helpers_test で mixin される側の
モジュール契約（メソッドの有無・constant）を固定する。

注: App の Metrics/ClassLength disable は外せなかった。helper 抽出後も route 本体
だけで 175 行（>130）あるため。route を別 Sinatra クラス / extension へ割れば
外せるが、request テストの土台が無く回帰を検知できない（Database の migration
抽出を見送ったのと同じ理由・#27 スコープ外）。土台が入ってから別途。
